refactor(client): move FileManager & update
